Corrected id data type in repositories

+++ src/main/java/com/kozik/nursery/repositories/AddressRepository.java
@@ -4,6 +4,6 @@ import com.kozik.nursery.entities.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface AddressRepository extends JpaRepository<Address, Integer>{
+public interface AddressRepository extends JpaRepository<Address, Long>{
     
 }
